requirements.txt
python-dotenv
certifi
pymongo
fastapi
uvicorn[standard]
motor
fastapi[standard]
passlib[bcrypt]
python-jose